Reorganize files in /src
