--- metaspace/webapp/src/about.md
@@ -1,50 +1,0 @@
-![METASPACE here and now](./assets/METASPACE_banner_blue_bg.jpg)
-
-## METASPACE provides an open platform for metabolite imaging mass spectrometry
-
-METASPACE platform hosts an engine for metabolite annotation of imaging mass spectrometry data
-as well as a spatial metabolite knowledgebase of the metabolites from hundreds of public datasets
-provided by the community.
-
-METASPACE platform is developed by software engineers, data scientists and mass spectrometrists
-from the [Alexandrov team at EMBL](https://www.embl.de/research/units/scb/alexandrov/).
-This work is a part of the [European project METASPACE](https://metaspace2020.eu).
-
-## Open as public space
-
-We do not hide anything. All code is open-source, the input format is the [imzML](http://imzml.org/)
-supported by all mass spec major vendors, the metabolite annotations from the community datasets are public
-and can be browsed or exported.
-
-## How can I contribute?
-
-Please read [our publication in Nature Methods](https://www.nature.com/nmeth/journal/v14/n1/full/nmeth.4072.html),
-follow us on [Twitter](https://twitter.com/metaspace2020), spread the word, and think of us while doing good deeds :)
-
-- **Have high-resolution imaging mass spectrometry data**? Please consider uploading them.
-- **Are you a data scientist**? Please feel free to the export the metabolite annotations to mine thousands of metabolites found in hundreds of community-provided datasets.
-
-
-
-## How shall I cite METASPACE in a publication?
-
-Please cite our publication [(Palmer et al., 2016, Nature Methods)](http://www.nature.com/nmeth/journal/v14/n1/full/nmeth.4072.html) and refer to [https://metaspace2020.eu](https://metaspace2020.eu).
-
-## Please appreciate those who funded it
-
-<div style="display:flex;">
-  <div style="padding-right: 20px;">
-    <img src="./assets/eu_flag.jpg" alt="EU" style="height:60px;"></img>
-  </div>
-  <div style="display:flex; flex-direction:column; justify-content: center;">
-    This website is part of a project that has received funding from the European Union's Horizon 2020 research and innovation programme under grant agreement No 634402.
-  </div>
-</div>
-
-## How to stay in touch?
-
-Follow us on [Twitter](https://twitter.com/metaspace2020) or drop a line to [our email](mailto:contact@metaspace2020.eu), we will add you to our special list :)
-
-For all other inquiries or questions, please [contact us over email](mailto:contact@metaspace2020.eu).
-
-Cheers!
