+++ metaspace/webapp/src/modules/Account/components/SignInDialog.spec.ts
@@ -7,7 +7,7 @@ import account from '../store/account'
 import router from '../../../router'
 import { restoreConsole, suppressConsoleWarn } from '../../../../tests/utils/suppressConsole'
 import * as _mockAuthApi from '../../../api/auth'
-import { refreshLoginStatus as _mockRefreshLoginStatus } from '../../../graphqlClient'
+import { refreshLoginStatus as _mockRefreshLoginStatus } from '../../../api/graphqlClient'
 
 jest.mock('../../../api/auth')
 const mockAuthApi = _mockAuthApi as jest.Mocked<typeof _mockAuthApi>
