+++ metaspace/webapp/src/modules/Annotations/AnnotationView.ts
@@ -1,4 +1,4 @@
-import { renderMolFormulaHtml } from '../../util'
+import { renderMolFormulaHtml } from '../../lib/util'
 import DatasetInfo from '../../components/DatasetInfo.vue'
 import ColocalizationSettings from './annotation-widgets/ColocalizationSettings.vue'
 import { annotationQuery } from '../../api/annotation'
@@ -19,7 +19,7 @@ import { currentUserRoleQuery, CurrentUserRoleResult } from '../../api/user'
 import safeJsonParse from '../../lib/safeJsonParse'
 import { omit, pick, sortBy, throttle } from 'lodash-es'
 import { ANNOTATION_SPECIFIC_FILTERS } from '../Filters/filterSpecs'
-import config from '../../config'
+import config from '../../lib/config'
 import { OpacityMode } from '../../lib/createColormap'
 import CandidateMoleculesPopover from './annotation-widgets/CandidateMoleculesPopover.vue'
 import RelatedMolecules from './annotation-widgets/RelatedMolecules.vue'
