+++ metaspace/webapp/src/modules/Annotations/annotation-widgets/CandidateMoleculesPopover.vue
@@ -53,7 +53,7 @@
 </template>
 
 <script>
-import config from '../../../config'
+import config from '../../../lib/config'
 
 export default {
   inheritAttrs: false,
