+++ metaspace/webapp/src/modules/Annotations/annotation-widgets/default/DiagnosticsMetrics.vue
@@ -31,7 +31,7 @@
 import Vue from 'vue'
 import { Component, Prop } from 'vue-property-decorator'
 
-import config from '../../../../config'
+import config from '../../../../lib/config'
 
 @Component({
 })
