+++ metaspace/webapp/src/modules/App/NewFeaturePopup.vue
@@ -48,7 +48,7 @@
 import Vue from 'vue'
 import Popper, { Placement } from 'popper.js'
 
-import config from '../../config'
+import config from '../../lib/config'
 import Component from 'vue-class-component'
 import { debounce, isArray } from 'lodash-es'
 import { getLocalStorage, setLocalStorage } from '../../lib/localStorage'
