+++ metaspace/webapp/src/modules/App/element-variables.scss
