+++ metaspace/webapp/src/modules/Datasets/list/DatasetThumbnail.vue
@@ -67,7 +67,7 @@
 </template>
 
 <script>
-import config from '../../../config'
+import config from '../../../lib/config'
 export default {
   props: ['dataset', 'editable'],
   data() {
