+++ metaspace/webapp/src/modules/MetadataEditor/UploadPage.vue
@@ -98,8 +98,8 @@ import FineUploader from './inputs/FineUploader.vue'
 import MetadataEditor from './MetadataEditor.vue'
 import Vue from 'vue'
 
-import config from '../../config'
-import { pathFromUUID } from '../../util'
+import config from '../../lib/config'
+import { pathFromUUID } from '../../lib/util'
 import { createDatasetQuery } from '../../api/dataset'
 import { getSystemHealthQuery, getSystemHealthSubscribeToMore } from '../../api/system'
 import get from 'lodash-es/get'
