+++ metaspace/webapp/src/types/dom-to-image-google-font-issue.d.ts
