+++ metaspace/webapp/tests/utils/setupTestFramework.ts
@@ -4,7 +4,7 @@ import registerMockComponent from './registerMockComponent'
 import VueRouter from 'vue-router'
 import registerMockDirective from './registerMockDirective'
 import { Wrapper, config as vueTestConfig } from '@vue/test-utils'
-import { replaceConfigWithDefaultForTests } from '../../src/config'
+import { replaceConfigWithDefaultForTests } from '../../src/lib/config'
 import VueCompositionApi from '@vue/composition-api'
 
 window.fetch = jest.fn()
