+++ metaspace/webapp/tsconfig.json
@@ -18,7 +18,7 @@
     "jsx": "preserve"
   },
   "include": [
-    "src/vue-shims.d.ts",
+    "src/types/*.d.ts",
     "src/**/*.ts",
     "src/**/*.tsx",
     "src/**/*.vue"
